ENH: Defer upgrade of idc-index until Slicer is restarted
When the module is started, it runs a check in the background to see if the package is outdated.
If it is, then it sets a flag that will update the package the next time Slicer is started.

A message will also be displayed at the top of the module to let the user know that an update is available.

Re #12