ENH: Dock the IDC browser widget into the main layout
Previously the browser widget was a floating window separate from the main window. Now the browser window is incorporated into the main layout.
When entering the module, the browser will be "opened" by changing the layout to one that includes the IDC browser widget. Exiting the module will restore the layout.
This behavior is consistent with the DICOM browser widget.

Re #50